better dataflow algorithm

+++ libafl_targets/src/sancov_pcguard.rs
@@ -45,7 +45,7 @@ pub unsafe extern "C" fn __sanitizer_cov_trace_pc_guard(guard: *mut u32) {
     }
     #[cfg(feature = "dataflow")]
     {
-        crate::dataflow::LAST_GUARD = *guard as usize;
+        crate::dataflow::LAST_GUARD = *guard;
     }
 }
